Add retry for token exchange + improve tests
