+++ security/pkg/nodeagent/plugin/plugin.go
@@ -26,5 +26,5 @@ const (
 
 // Plugin provides common interfaces so that authentication providers could choose to implement their specific logic.
 type Plugin interface {
-	ExchangeToken(context.Context, string, string) (string, time.Time, error)
+	ExchangeToken(context.Context, string, string) (string, time.Time, int, error)
 }
